updating test files

+++ isis/src/base/objs/UserInterface/UserInterface.cpp
@@ -394,7 +394,7 @@ namespace Isis {
    * This is used to load the command line into p_cmdline and the Aml object
    * using information contained in argc and argv.
    *
-   * @param args QVector of arguments 
+   * @param args QVector of arguments
    *
    * @throws Isis::IException::User - Invalid value for reserve parameter
    * @throws Isis::IException::User - Invalid command line
@@ -405,22 +405,22 @@ namespace Isis {
    */
   void UserInterface::loadCommandLine(QVector<QString> &args, bool ignoreAppName) {
     char **c_args;
-    
+
     if (ignoreAppName) {
       args.prepend("someapp");
     }
 
     c_args = (char**)malloc(sizeof(char*)*args.size());
-    
-    for (size_t i = 0; i < args.size(); i++) {
+
+    for (int i = 0; i < args.size(); i++) {
       c_args[i] = (char*)malloc(sizeof(char)*args[i].size());
       strcpy(c_args[i], args[i].toLatin1().data());
     }
 
     loadCommandLine(args.size(), c_args);
-  }  
-    
-    
+  }
+
+
   /**
    * This is used to load the command line into p_cmdline and the Aml object
    * using information contained in argc and argv.
